move performTap under Driver
